Updated to 1.6.5.1

+++ src/Pro/Admin/Entries/Export/File.php
@@ -2,8 +2,7 @@
 
 namespace WPForms\Pro\Admin\Entries\Export;
 
-use \Goodby\CSV\Export\Standard\Exporter;
-use \Goodby\CSV\Export\Standard\ExporterConfig;
+use Goodby\CSV\Export\Standard\CsvFileObject;
 
 /**
  * File-related routines.
@@ -42,9 +41,9 @@ class File {
 	 */
 	public function hooks() {
 
-		add_action( 'wpforms_tools_init', array( $this, 'entries_export_download_file' ) );
-		add_action( 'wpforms_tools_init', array( $this, 'single_entry_export_download_file' ) );
-		add_action( 'admin_init', array( $this, 'remove_old_export_files' ) );
+		add_action( 'wpforms_tools_init', [ $this, 'entries_export_download_file' ] );
+		add_action( 'wpforms_tools_init', [ $this, 'single_entry_export_download_file' ] );
+		add_action( Export::TASK_CLEANUP, [ $this, 'remove_old_export_files' ] );
 	}
 
 	/**
@@ -52,10 +51,9 @@ class File {
 	 *
 	 * @since 1.5.5
 	 *
-	 * @param array $export_data  Export data array.
 	 * @param array $request_data Request data array.
 	 */
-	public function write_csv( $export_data, $request_data ) {
+	public function write_csv( $request_data ) {
 
 		$export_file = $this->get_tmpfname( $request_data );
 
@@ -63,14 +61,65 @@ class File {
 			return;
 		}
 
-		// Include Exporter.
-		require_once WPFORMS_PLUGIN_DIR . 'vendor/autoload.php';
+		$csv       = new CsvFileObject( $export_file, 'a' );
+		$enclosure = '"';
 
-		$config = new ExporterConfig();
-		$config->setDelimiter( $this->export->configuration['csv_export_separator'] );
-		$config->setFileMode( 'a' );
-		$exporter = new Exporter( $config );
-		$exporter->export( $export_file, $export_data );
+		$csv->fputcsv( $request_data['columns_row'], $this->export->configuration['csv_export_separator'], $enclosure );
+
+		for ( $i = 1; $i <= $request_data['total_steps']; $i ++ ) {
+
+			$entries = wpforms()->entry->get_entries( $request_data['db_args'] );
+
+			foreach ( $this->export->ajax->get_entry_data( $entries ) as $entry ) {
+
+				$csv->fputcsv( $entry, $this->export->configuration['csv_export_separator'], $enclosure );
+			}
+
+			$request_data['db_args']['offset'] = $i * $this->export->configuration['entries_per_step'];
+		}
+
+		$csv->fflush();
+	}
+
+	/**
+	 * Write data to .xlsx file.
+	 *
+	 * @since 1.6.5
+	 *
+	 * @param array $request_data Request data array.
+	 */
+	public function write_xlsx( $request_data ) {
+
+		$export_file = $this->get_tmpfname( $request_data );
+
+		if ( empty( $export_file ) ) {
+			return;
+		}
+
+		$writer     = new \XLSXWriter();
+		$sheet_name = 'WPForms';
+
+		$widths = [];
+
+		foreach ( $request_data['columns_row'] as $header ) {
+			$widths[] = strlen( $header ) + 2;
+		}
+
+		$writer->writeSheetHeader( $sheet_name, array_fill_keys( $request_data['columns_row'], 'string' ), [ 'widths' => $widths ] );
+
+		for ( $i = 1; $i <= $request_data['total_steps']; $i ++ ) {
+
+			$entries = wpforms()->entry->get_entries( $request_data['db_args'] );
+
+			foreach ( $this->export->ajax->get_entry_data( $entries ) as $entry ) {
+
+				$writer->writeSheetRow( $sheet_name, $entry, [ 'wrap_text' => true ] );
+			}
+
+			$request_data['db_args']['offset'] = $i * $this->export->configuration['entries_per_step'];
+		}
+
+		$writer->writeToFile( $export_file );
 	}
 
 	/**
@@ -172,10 +221,14 @@ class File {
 
 		$entry_suffix = ! empty( $request_data['db_args']['entry_id'] ) ? '-entry-' . $request_data['db_args']['entry_id'] : '';
 
-		$file_name = 'wpforms-' . $request_data['db_args']['form_id'] . '-' . sanitize_file_name( get_the_title( $request_data['db_args']['form_id'] ) ) . $entry_suffix . '-' . current_time( 'Y-m-d-H-i-s' ) . '.csv';
+		$file_name = 'wpforms-' . $request_data['db_args']['form_id'] . '-' . sanitize_file_name( get_the_title( $request_data['db_args']['form_id'] ) ) . $entry_suffix . '-' . current_time( 'Y-m-d-H-i-s' ) . '.' . $request_data['type'];
+
 		$this->http_headers( $file_name );
 
 		readfile( $export_file ); // phpcs:ignore
+
+		// Schedule clean up.
+		$this->schedule_remove_old_export_files();
 
 		exit;
 	}
@@ -282,11 +335,11 @@ class File {
 
 			$this->export->ajax->request_data = $request_data;
 
-			// Get export data.
-			$export_data = $this->export->ajax->get_data();
-
-			// Writing to csv file.
-			$this->write_csv( $export_data, $request_data );
+			if ( empty( $request_data['type'] ) || $request_data['type'] === 'csv' ) {
+				$this->write_csv( $request_data );
+			} elseif ( $request_data['type'] === 'xlsx' ) {
+				$this->write_xlsx( $request_data );
+			}
 
 			$this->output_file( $request_data );
 
@@ -302,19 +355,34 @@ class File {
 	}
 
 	/**
-	 * Garbage clearing.
-	 * Actually we need to clear only temporary files
-	 * because transients clears automagically.
+	 * Register a cleanup task to remove temp export files.
 	 *
-	 * TODO: rewrite to use wp-cron or actions scheduler.
+	 * @since 1.6.5
+	 */
+	public function schedule_remove_old_export_files() {
+
+		$tasks = wpforms()->get( 'tasks' );
+
+		if ( ! empty( $tasks->is_scheduled( Export::TASK_CLEANUP ) ) ) {
+			return;
+		}
+
+		$tasks->create( Export::TASK_CLEANUP )
+		      ->recurring( time() + 60, $this->export->configuration['request_data_ttl'] )
+		      ->params()
+		      ->register();
+	}
+
+	/**
+	 * Garbage clearing.
+	 * We need to clear only temporary files because transients are cleared automagically.
 	 *
 	 * @since 1.5.5
+	 * @since 1.6.5 Stop using transients. Now we use ActionScheduler for this task.
 	 */
 	public function remove_old_export_files() {
 
-		if ( (bool) get_transient( 'wpforms_entries_export_tmpdata_cleared' ) ) {
-			return;
-		}
+		clearstatcache();
 
 		$files = glob( $this->get_tmpdir() . '/*' );
 		$now   = time();
@@ -322,12 +390,11 @@ class File {
 		foreach ( $files as $file ) {
 			if (
 				is_file( $file ) &&
+				pathinfo( $file, PATHINFO_BASENAME ) !== 'index.html' &&
 				( $now - filemtime( $file ) ) > $this->export->configuration['request_data_ttl']
 			) {
 				unlink( $file );
 			}
 		}
-
-		set_transient( 'wpforms_entries_export_tmpdata_cleared', 'yes', HOUR_IN_SECONDS );
 	}
 }
